requirements.txt
tensorflow==1.15.5
protobuf==3.20.3
pillow
scipy
matplotlib
keras
fastapi 
uvicorn 
numpy==1.16.0
